+++ CoreGraphFramework/src/au/gov/asd/tac/constellation/graph/schema/visual/GraphLabel.java
@@ -129,7 +129,7 @@ public final class GraphLabel implements Serializable, Comparable<GraphLabel> {
             final String attributeName = labelProperties.get(0);
             final ConstellationColor color = ConstellationColor.getColorValue(labelProperties.get(1));
             if (color == null) {
-                throw new IllegalArgumentException("Undefined colour for label.");
+                throw new IllegalArgumentException("Undefined color for label.");
             }
             final float size;
             try {