+++ CoreLayersView/src/au/gov/asd/tac/constellation/views/layers/utilities/UpdateElementBitmaskPlugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2021 Australian Signals Directorate
+ * Copyright 2010-2022 Australian Signals Directorate
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -57,9 +57,9 @@ public final class UpdateElementBitmaskPlugin extends SimpleEditPlugin {
         for (int vertexPosition = 0; vertexPosition < vertexCount; vertexPosition++) {
             final int vertexId = graph.getVertex(vertexPosition);
             if (!selectedOnly || graph.getBooleanValue(vertexSelectedAttributeId, vertexId)) {
-                if (layerAction.equals(LayerAction.ADD)) {
+                if (layerAction == LayerAction.ADD) {
                     graph.setLongValue(vertexBitmaskAttributeId, vertexId, graph.getLongValue(vertexBitmaskAttributeId, vertexId) | (1 << targetMask));
-                } else if (layerAction.equals(LayerAction.REMOVE)) {
+                } else if (layerAction == LayerAction.REMOVE) {
                     graph.setLongValue(vertexBitmaskAttributeId, vertexId, graph.getLongValue(vertexBitmaskAttributeId, vertexId) & ~(1 << targetMask));
                 } else {
                     // Do nothing
@@ -75,9 +75,9 @@ public final class UpdateElementBitmaskPlugin extends SimpleEditPlugin {
         for (int transactionPosition = 0; transactionPosition < transactionCount; transactionPosition++) {
             final int transactionId = graph.getTransaction(transactionPosition);
             if (!selectedOnly || graph.getBooleanValue(transactionSelectedAttributeId, transactionId)) {
-                if (layerAction.equals(LayerAction.ADD)) {
+                if (layerAction == LayerAction.ADD) {
                     graph.setLongValue(transactionBitmaskAttributeId, transactionId, graph.getLongValue(transactionBitmaskAttributeId, transactionId) | (1 << targetMask));
-                } else if (layerAction.equals(LayerAction.REMOVE)) {
+                } else if (layerAction == LayerAction.REMOVE) {
                     graph.setLongValue(transactionBitmaskAttributeId, transactionId, graph.getLongValue(transactionBitmaskAttributeId, transactionId) & ~(1 << targetMask));
                 } else {
                     // Do nothing